// client-app/src/app/Shared/Components/Client/shared-client.ts
export class SharedClient {
    clientID: any;
    firstName: string;
    lastName: string;
    dob: any;
    genderID: any;
    raceID: any;
    address: string;
    cityID: any;
    stateID: any;
    zipcodeID: any;
    countyID: any;
    ssn: string;
    ethnicityID: any;
    homePh: any;
    fax: any;
    email: string;
    workPh: any;
    cellPh: any;
    mi: any;
    suffix: any;
    tribeID: any;
    insuranceTypeID: any;
    medicaid: any;
    isNotLiveBirth: boolean;
    kaecses: number;
    kees: any;
    alternateID: any;
    notes: any;
    acronym: boolean;
    nonContractClient: boolean;
    conST: any;
    medicalID: any;
    kanCareMCO_ProviderID: any;
    facts: any;
    entity: string ;
}